REFACTOR: appcontext -> server
