+++ core/internal/pkg/server/server_setup.go
@@ -1,4 +1,4 @@
-package appcontext
+package server
 
 import (
 	"context"
